some fixes, refactor settings to kotlin
