+++ MessageLogger/src/main/java/io/github/juby210/acplugins/messagelogger/MessageRecord.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Juby210
+ * Copyright (c) 2021-2022 Juby210
  * Licensed under the Open Software License version 3.0
  */
